Refactor RadarEntry
